Move GetStats to the correct package
